Moved logging unit tests to be together with future viewer tests, to keep structure consistent
