+++ caQtDM_Tests/unitTests/tst_viewer/tst_logging/tst_generalloghandler.cpp
