+++ caQtDM_Tests/unitTests/tst_viewer/tst_logging/tst_logstashloghandler.h
