+++ caQtDM_Tests/unitTests/unitTests.pro
@@ -1,3 +1,3 @@
 TEMPLATE = subdirs
 
-SUBDIRS += tst_logging tst_qtcontrols tst_lib tst_plugins
+SUBDIRS += tst_viewer tst_qtcontrols tst_lib tst_plugins
